requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dbt-core
dbt-postgres
apache-airflow-providers-postgres
apache-airflow-providers-http